Moved LenientDateTypeAdapter to own class in new util package
